refactor(editor): Migrate NDV settings and runData to workflowDocumentStore
Replace direct workflowsStore/workflowState node access with injectWorkflowDocumentStore() in NDV components and useWorkflowDocumentStore() computed pattern in composables. This decouples node document state from the monolithic workflows store.

Files migrated:
- NodeSettings.vue: inject pattern for component
- useNodeSettingsParameters.ts: computed pattern for composable
- ParameterInput.vue: inject pattern for component
- RunData.vue: inject pattern for component
- VirtualSchema.vue: inject pattern for component
- SQLEditor.test.ts: mock injectWorkflowDocumentStore

All tests pass (147 passed). Related to CAT-2518.
